Migrate to typescript sheet builder
